Проект поделен на модули
